+++ src/new-components/heading/heading.stories.tsx
@@ -17,51 +17,18 @@ export function HeadingStory() {
         <section className="story">
             <Stack dividers="primary" space="medium">
                 <Stack space="small">
-                    <Heading level={1} size="largest">
-                        Heading level 1, largest
-                    </Heading>
-                    <Heading level={1} size="larger">
-                        Heading level 1, larger
-                    </Heading>
                     <Heading level={1}>Heading level 1</Heading>
-                    <Heading level={1} size="smaller">
-                        Heading level 1, smaller
-                    </Heading>
                 </Stack>
 
                 <Stack space="small">
-                    <Heading level={2} size="largest">
-                        Heading level 2, largest
-                    </Heading>
-                    <Heading level={2} size="larger">
-                        Heading level 2, larger
-                    </Heading>
                     <Heading level={2}>Heading level 2</Heading>
-                    <Heading level={2} size="smaller">
-                        Heading level 2, smaller
-                    </Heading>
                 </Stack>
 
                 <Stack space="small">
-                    <Heading level={3} size="largest">
-                        Heading level 3, largest
-                    </Heading>
-                    <Heading level={3} size="larger">
-                        Heading level 3, larger
-                    </Heading>
                     <Heading level={3}>Heading level 3</Heading>
-                    <Heading level={3} size="smaller">
-                        Heading level 3, smaller
-                    </Heading>
                 </Stack>
 
                 <Stack space="small">
-                    <Heading level={4} size="largest">
-                        Heading level 4 / 5 / 6, largest
-                    </Heading>
-                    <Heading level={4} size="larger">
-                        Heading level 4 / 5 / 6, larger
-                    </Heading>
                     <Heading level={4}>Heading level 4 / 5 / 6</Heading>
                 </Stack>
             </Stack>
@@ -72,7 +39,7 @@ export function HeadingStory() {
 export function TruncatedHeadingStory() {
     return (
         <section className="story">
-            <Heading level={1} size="largest" lineClamp={1}>
+            <Heading level={1} lineClamp={1}>
                 This is a long title which we will use demonstrate truncating content. When this
                 overflows and begins to drop to a new line, its overflowing content will be replaced
                 by ellipses.